# Optimize dispatch for non `__torch_function__` calls