--- a/aten/src/ATen/core/interned_strings.h
+++ b/aten/src/ATen/core/interned_strings.h
@@ -284,6 +284,8 @@ namespace c10 {
   _(aten, swapaxes_)                 \
   _(aten, swapdims)                  \
   _(aten, swapdims_)                 \
+  _(aten, has_torch_function)        \
+  _(aten, object_has_torch_function) \
   FORALL_ATEN_BASE_SYMBOLS(_)        \
   _(onnx, Add)                       \
   _(onnx, Concat)                    \