--- a/torch/csrc/autograd/python_variable.cpp
+++ b/torch/csrc/autograd/python_variable.cpp
@@ -43,6 +43,7 @@ using namespace torch::autograd;
 namespace py = pybind11;
 
 PyObject *THPVariableClass = nullptr;
+PyObject *ParameterClass = nullptr;
 
 static const char* VOLATILE_WARNING =
     "volatile was removed and now has no effect. Use "