--- a/torch/csrc/autograd/python_variable.h
+++ b/torch/csrc/autograd/python_variable.h
@@ -19,12 +19,20 @@ struct THPVariable {
 };
 
 THP_API PyObject *THPVariableClass;
+THP_API PyObject *ParameterClass;
 
 bool THPVariable_initModule(PyObject *module);
 THP_API PyObject * THPVariable_Wrap(torch::autograd::Variable var);
 
 static inline bool THPVariable_CheckExact(PyObject *obj) {
   return Py_TYPE(obj) == (PyTypeObject*)THPVariableClass;
+}
+
+static inline bool THPVariableOrParameter_CheckExact(PyObject *obj) {
+  auto py_type = Py_TYPE(obj);
+  return (
+    py_type == (PyTypeObject*)THPVariableClass ||
+    py_type == (PyTypeObject*)ParameterClass);
 }
 
 inline bool THPVariable_Check(PyObject *obj)