--- a/torch/csrc/utils/disable_torch_function.h
+++ b/torch/csrc/utils/disable_torch_function.h
@@ -9,8 +9,11 @@ namespace torch {
   bool torch_function_enabled();
   PyObject* disabled_torch_function_impl();
   void set_disabled_torch_function_impl(PyObject* value);
+  bool fast_has_torch_function(PyObject *obj);
 }
 
 PyObject* THPModule_isEnabledTorchFunction(PyObject* self, PyObject *unused);
 PyObject* THPModule_DisableTorchFunctionType();
 PyObject* THPModule_disable_torch_function(PyObject *self, PyObject *args);
+PyObject* THPModule_object_has_torch_function(PyObject*, PyObject *obj);
+PyObject* THPModule_has_torch_function(PyObject*, PyObject *arg);