--- a/torch/functional.py
+++ b/torch/functional.py
@@ -5,6 +5,7 @@ from typing import (
 import torch
 import torch.nn.functional as F
 from torch.types import _size
+from torch._C import _is_tensor
 from ._lowrank import svd_lowrank, pca_lowrank
 from .overrides import has_torch_function, handle_torch_function
 from ._jit_internal import boolean_dispatch, List
@@ -67,7 +68,7 @@ def broadcast_tensors(*tensors):
                 [0, 1, 2]])
     """
     if not torch.jit.is_scripting():
-        if any(type(t) is not Tensor for t in tensors) and has_torch_function(tensors):
+        if has_torch_function(tensors):
             return handle_torch_function(broadcast_tensors, tensors, *tensors)
     return _VF.broadcast_tensors(tensors)  # type: ignore
 
@@ -112,7 +113,7 @@ def split(tensor, split_size_or_sections, dim=0):
                  [8, 9]]))
     """
     if not torch.jit.is_scripting():
-        if type(tensor) is not Tensor and has_torch_function((tensor,)):
+        if not _is_tensor(tensor) and has_torch_function((tensor,)):
             return handle_torch_function(split, (tensor,), tensor, split_size_or_sections,
                                          dim=dim)
     # Overwriting reason:
@@ -202,7 +203,7 @@ def lu_unpack(LU_data, LU_pivots, unpack_data=True, unpack_pivots=True):
     """
     if not torch.jit.is_scripting():
         tens_ops = (LU_data, LU_pivots)
-        if any([type(t) is not Tensor for t in tens_ops]) and has_torch_function(tens_ops):
+        if has_torch_function(tens_ops):
             return handle_torch_function(
                 lu_unpack, tens_ops, LU_data, LU_pivots, unpack_data=unpack_data,
                 unpack_pivots=unpack_pivots)
@@ -333,7 +334,7 @@ Examples::
     torch.Size([2, 3, 5, 4])
 """
     if not torch.jit.is_scripting():
-        if any(type(t) is not Tensor for t in operands) and has_torch_function(operands):
+        if has_torch_function(operands):
             return handle_torch_function(einsum, operands, equation, *operands)
     if len(operands) == 1 and isinstance(operands[0], (list, tuple)):
         # the old interface of passing the operands as one list argument
@@ -383,7 +384,7 @@ else:
 
 def _meshgrid(*tensors):
     if not torch.jit.is_scripting():
-        if any(type(t) is not Tensor for t in tensors) and has_torch_function(tensors):
+        if has_torch_function(tensors):
             return handle_torch_function(meshgrid, tensors, *tensors)
     if len(tensors) == 1 and isinstance(tensors[0], (list, tuple)):
         # the old interface of passing the operands as one list argument
@@ -499,7 +500,7 @@ def stft(input: Tensor, n_fft: int, hop_length: Optional[int] = None,
 
     """
     if not torch.jit.is_scripting():
-        if type(input) is not Tensor and has_torch_function((input,)):
+        if not _is_tensor(input) and has_torch_function((input,)):
             return handle_torch_function(
                 stft, (input,), input, n_fft, hop_length=hop_length, win_length=win_length,
                 window=window, center=center, pad_mode=pad_mode, normalized=normalized,
@@ -577,7 +578,7 @@ def istft(input: Tensor, n_fft: int, hop_length: Optional[int] = None,
         Tensor: Least squares estimation of the original signal of size (..., signal_length)
     """
     if not torch.jit.is_scripting():
-        if type(input) is not Tensor and has_torch_function((input,)):
+        if not _is_tensor(input) and has_torch_function((input,)):
             return handle_torch_function(
                 istft, (input,), input, n_fft, hop_length=hop_length, win_length=win_length,
                 window=window, center=center, normalized=normalized, onesided=onesided,
@@ -661,7 +662,7 @@ def _unique_impl(input: Tensor, sorted: bool = True,
 
     """
     if not torch.jit.is_scripting():
-        if type(input) is not Tensor and has_torch_function((input,)):
+        if not _is_tensor(input) and has_torch_function((input,)):
             return handle_torch_function(
                 unique, (input,), input, sorted=sorted, return_inverse=return_inverse,
                 return_counts=return_counts, dim=dim)
@@ -737,7 +738,7 @@ def _unique_consecutive_impl(input: Tensor, return_inverse: bool = False,
         tensor([2, 2, 1, 2, 1])
     """
     if not torch.jit.is_scripting():
-        if type(input) is not Tensor and has_torch_function((input,)):
+        if not _is_tensor(input) and has_torch_function((input,)):
             return handle_torch_function(
                 unique_consecutive, (input,), input, return_inverse=return_inverse,
                 return_counts=return_counts, dim=dim)
@@ -750,7 +751,7 @@ def _return_counts(input, sorted=True, return_inverse=False, return_counts=False
     # type: (Tensor, bool, bool, bool, Optional[int]) -> Tuple[Tensor, Tensor]
 
     if not torch.jit.is_scripting():
-        if type(input) is not Tensor and has_torch_function((input,)):
+        if not _is_tensor(input) and has_torch_function((input,)):
             return _unique_impl(input, sorted, return_inverse, return_counts, dim)
 
     output, _, counts = _unique_impl(input, sorted, return_inverse, return_counts, dim)
@@ -761,7 +762,7 @@ def _return_output(input, sorted=True, return_inverse=False, return_counts=False
     # type: (Tensor, bool, bool, bool, Optional[int]) -> Tensor
 
     if not torch.jit.is_scripting():
-        if type(input) is not Tensor and has_torch_function((input,)):
+        if not _is_tensor(input) and has_torch_function((input,)):
             return _unique_impl(input, sorted, return_inverse, return_counts, dim)
 
     output, _, _ = _unique_impl(input, sorted, return_inverse, return_counts, dim)
@@ -772,7 +773,7 @@ def _return_inverse(input, sorted=True, return_inverse=False, return_counts=Fals
     # type: (Tensor, bool, bool, bool, Optional[int]) -> Tuple[Tensor, Tensor]
 
     if not torch.jit.is_scripting():
-        if type(input) is not Tensor and has_torch_function((input,)):
+        if not _is_tensor(input) and has_torch_function((input,)):
             return _unique_impl(input, sorted, return_inverse, return_counts, dim)
 
     output, inverse_indices, _ = _unique_impl(input, sorted, return_inverse, return_counts, dim)
@@ -815,7 +816,7 @@ def _consecutive_return_counts(input, return_inverse=False, return_counts=False,
     # type: (Tensor, bool, bool, Optional[int]) -> Tuple[Tensor, Tensor]
 
     if not torch.jit.is_scripting():
-        if type(input) is not Tensor and has_torch_function((input,)):
+        if not _is_tensor(input) and has_torch_function((input,)):
             return _unique_consecutive_impl(input, return_inverse, return_counts, dim)
 
     output, _, counts = _unique_consecutive_impl(input, return_inverse, return_counts, dim)
@@ -826,7 +827,7 @@ def _consecutive_return_output(input, return_inverse=False, return_counts=False,
     # type: (Tensor, bool, bool, Optional[int]) -> Tensor
 
     if not torch.jit.is_scripting():
-        if type(input) is not Tensor and has_torch_function((input,)):
+        if not _is_tensor(input) and has_torch_function((input,)):
             return _unique_consecutive_impl(input, return_inverse, return_counts, dim)
 
     output, _, _ = _unique_consecutive_impl(input, return_inverse, return_counts, dim)
@@ -837,7 +838,7 @@ def _consecutive_return_inverse(input, return_inverse=False, return_counts=False
     # type: (Tensor, bool, bool, Optional[int]) -> Tuple[Tensor, Tensor]
 
     if not torch.jit.is_scripting():
-        if type(input) is not Tensor and has_torch_function((input,)):
+        if not _is_tensor(input) and has_torch_function((input,)):
             return _unique_consecutive_impl(input, return_inverse, return_counts, dim)
 
     output, inverse_indices, _ = _unique_consecutive_impl(input, return_inverse, return_counts, dim)
@@ -927,7 +928,7 @@ def tensordot(a, b, dims=2, out=None):
                 [ -0.2850,   4.2573,  -3.5997]])
     """
     if not torch.jit.is_scripting():
-        if (type(a) is not Tensor or type(b) is not Tensor) and has_torch_function((a, b)):
+        if has_torch_function((a, b)):
             return handle_torch_function(tensordot, (a, b), a, b, dims=dims)
     if isinstance(dims, (list, tuple)) or \
        (isinstance(dims, torch.Tensor) and dims.numel() > 1):
@@ -973,7 +974,7 @@ def cartesian_prod(*tensors):
                 [3, 5]])
     """
     if not torch.jit.is_scripting():
-        if any(type(t) is not Tensor for t in tensors) and has_torch_function(tensors):
+        if has_torch_function(tensors):
             return handle_torch_function(cartesian_prod, tensors, *tensors)
     return _VF.cartesian_prod(tensors)  # type: ignore
 
@@ -1007,7 +1008,7 @@ def block_diag(*tensors):
                 [0, 0, 0, 0, 0, 0, 0, 0, 0, 5],
                 [0, 0, 0, 0, 0, 0, 0, 0, 0, 6]])
     """
-    if any(type(t) is not Tensor for t in tensors) and has_torch_function(tensors):
+    if has_torch_function(tensors):
         return handle_torch_function(block_diag, tensors, *tensors)
     return torch._C._VariableFunctions.block_diag(tensors)  # type: ignore
 
@@ -1055,7 +1056,7 @@ def cdist(x1, x2, p=2., compute_mode='use_mm_for_euclid_dist_if_necessary'):
                 [2.2830, 0.3791]])
     """
     if not torch.jit.is_scripting():
-        if (type(x1) is not Tensor or type(x2) is not Tensor) and has_torch_function((x1, x2)):
+        if has_torch_function((x1, x2)):
             return handle_torch_function(
                 cdist, (x1, x2), x1, x2, p=p, compute_mode=compute_mode)
     if compute_mode == 'use_mm_for_euclid_dist_if_necessary':
@@ -1095,7 +1096,7 @@ def atleast_1d(*tensors):
         (tensor([0.5000]), tensor([1.]))
     """
     if not torch.jit.is_scripting():
-        if any(type(t) is not Tensor for t in tensors) and has_torch_function(tensors):
+        if has_torch_function(tensors):
             return handle_torch_function(atleast_1d, tensors, *tensors)
     if len(tensors) == 1:
         tensors = tensors[0]
@@ -1130,7 +1131,7 @@ def atleast_2d(*tensors):
         (tensor([[0.5000]]), tensor([[1.]]))
     """
     if not torch.jit.is_scripting():
-        if any(type(t) is not Tensor for t in tensors) and has_torch_function(tensors):
+        if has_torch_function(tensors):
             return handle_torch_function(atleast_2d, tensors, *tensors)
     if len(tensors) == 1:
         tensors = tensors[0]
@@ -1174,7 +1175,7 @@ def atleast_3d(*tensors):
         (tensor([[[0.5000]]]), tensor([[[1.]]]))
     """
     if not torch.jit.is_scripting():
-        if any(type(t) is not Tensor for t in tensors) and has_torch_function(tensors):
+        if has_torch_function(tensors):
             return handle_torch_function(atleast_3d, tensors, *tensors)
     if len(tensors) == 1:
         tensors = tensors[0]
@@ -1301,7 +1302,7 @@ def norm(input, p="fro", dim=None, keepdim=False, out=None, dtype=None):  # noqa
     """
 
     if not torch.jit.is_scripting():
-        if type(input) is not Tensor and has_torch_function((input,)):
+        if not _is_tensor(input) and has_torch_function((input,)):
             return handle_torch_function(
                 norm, (input,), input, p=p, dim=dim, keepdim=keepdim, out=out, dtype=dtype)
 
@@ -1397,7 +1398,7 @@ def chain_matmul(*matrices):
     .. _`[CLRS]`: https://mitpress.mit.edu/books/introduction-algorithms-third-edition
     """
     if not torch.jit.is_scripting():
-        if any(type(t) is not Tensor for t in matrices) and has_torch_function(matrices):
+        if has_torch_function(matrices):
             return handle_torch_function(chain_matmul, matrices, *matrices)
     return _VF.chain_matmul(matrices)  # type: ignore
 
@@ -1517,7 +1518,7 @@ def _check_list_size(out_len: int, get_infos: bool, out: _ListOrSeq) -> None:
 def _lu_with_infos(A, pivot=True, get_infos=False, out=None):
     # type: (Tensor, bool, bool, Optional[Tuple[Tensor, Tensor, Tensor]]) -> Tuple[Tensor, Tensor, Tensor]
     if not torch.jit.is_scripting():
-        if type(A) is not Tensor and has_torch_function((A,)):
+        if not _is_tensor(A) and has_torch_function((A,)):
             return handle_torch_function(
                 lu, (A,), A, pivot=pivot, get_infos=get_infos, out=out)
     result = _lu_impl(A, pivot, get_infos, out)
@@ -1533,7 +1534,7 @@ def _lu_no_infos(A, pivot=True, get_infos=False, out=None):
     # type: (Tensor, bool, bool, Optional[Tuple[Tensor, Tensor]]) -> Tuple[Tensor, Tensor]
     # need to check for torch_function here so that we exit if
     if not torch.jit.is_scripting():
-        if type(A) is not Tensor and has_torch_function((A,)):
+        if not _is_tensor(A) and has_torch_function((A,)):
             return handle_torch_function(
                 lu, (A,), A, pivot=pivot, get_infos=get_infos, out=out)
     result = _lu_impl(A, pivot, get_infos, out)