--- a/torch/jit/_script.py
+++ b/torch/jit/_script.py
@@ -32,6 +32,7 @@ from torch.jit._state import (
     _set_jit_function_cache,
     _set_jit_overload_cache,
 )
+from torch.overrides import has_torch_function, object_has_torch_function
 
 torch._C.ScriptMethod.graph_for = _graph_for  # type: ignore
 torch._C.ScriptFunction.graph_for = _graph_for  # type: ignore
@@ -1081,3 +1082,5 @@ def _unwrap_optional(x):
 
 _register_builtin(_unwrap_optional, "aten::_unwrap_optional")
 _register_builtin(_jit_internal.is_scripting, "aten::is_scripting")
+_register_builtin(has_torch_function, "aten::has_torch_function")
+_register_builtin(object_has_torch_function, "aten::object_has_torch_function")