@robieta 2020-12-06
I've updated this PR and essentially adopted the approach discussed above. I've moved the logic for `has_torch_function` entirely into C++, and added a special case `object_has_torch_function` which skips tuple creation (and unpacking) for the single argument case since it's so common. And then I simply registered both to the JIT as static `False` methods.

Overall performance looks good: [Updated benchmark script](https://gist.github.com/robieta/3bf55fcaf23c0be35f10e8599aa6d87f)
<img width="1166" alt="Screen Shot 2020-12-06 at 3 44 45 PM" src="https://user-images.githubusercontent.com/13089297/101296745-fb640880-37d9-11eb-9d40-c7ab757fcaf4.png">

(Added before and after https://github.com/pytorch/pytorch/pull/37091 as well.) I'm still not thrilled with `Module` performance. The large difference in overhead for the later rows is due to passing `Parameter` rather than `Tensor` into `matmul`, so something is wonky on the binding front. (Maybe the Python arg parser.) Outside the scope of this PR, but definitely due for a follow up PR.