@robieta 2020-12-07
Also, I just realized I didn't add entries in the `.pyi` file. I'll do that when I fix the lint errors.