@robieta 2020-12-07
So apparently we are codegen-ing:

```
if(check_has_torch_function(self_))
```

rather than
```
if(!THPVariable_CheckExact(self_) && check_has_torch_function(self_))
```

(or `THPVariableOrParameter_CheckExact` / whatever we wind up calling the version that treats Parameters as Tensors) for no-arg ops (e.g. `x.t()`), so `Tensor` is going through the slower `check_has_torch_function`, and `Parameter` calls all the way back to Python.

For `.t()` it's about 0.3% instruction delta (50 ns, 2% wall time) on Tensor and 13% instructions (17% wall time) on Parameter.