@hameerabbasi 2020-12-07
> and `Parameter` calls all the way back to Python.

Could you clarify this? Technically, we check for exact `Tensor`, and we also check for `type.__torch_function__ is _disabled_torch_function_impl` attribute.