@robieta 2020-12-07
> Could you clarify this? Technically, we check for exact `Tensor`, and we also check for `type.__torch_function__` is `_disabled_torch_function_impl` attribute.

**python_variable_methods.cpp**
```
if(check_has_torch_function(self_)) {
  return handle_torch_function(self_, "t");
}
```

`check_has_torch_function` calls `PyTorch_LookupSpecial` which does an exact check, but allocates an extra `py::object` and does some pointer checks in the process. And `Parameter` goes through the last `PyObject_FastGetAttrString` path.