@hameerabbasi 2020-12-07
Apologies, pressed the wrong button.

`py::object` is a `PyObject*` essentially, so return `py::object()` is `return nullptr`.