@robieta 2020-12-07
> Apologies, pressed the wrong button.
> 
> `py::object` is a `PyObject*` essentially, so return `py::object()` is `return nullptr`.

I believe `py::handle` is the cheap one, and `py::object` also does ref counts. (It may be smart enough not to do it for the null case.) It's not a huge cost for Tensor, but it is slower than doing a type check at the start.