@ezyang 2020-12-07
> For .t() it's about 0.3% instruction delta (50 ns, 2% wall time) on Tensor and 13% instructions (450 ns, 17% wall time) on Parameter.

Maybe this is the smoking gun for the other CUDA problem?