@robieta 2020-12-07
> > For .t() it's about 0.3% instruction delta (50 ns, 2% wall time) on Tensor and 13% instructions (450 ns, 17% wall time) on Parameter.
> 
> Maybe this is the smoking gun for the other CUDA problem?

Possible, though I'm skeptical. I only noticed it because `Linear` calls `input.matmul(weight.t())`; however the CUDA issue is a ResNet and I don't see anything similar on conv2d, bn, or relu.